Extract generator support module
